test: Added course in ZIP file
This is a test to provide courses in a zip file through
GitHub pages. This may allow to bypass the GitHub APIs
limits and allow as to run courses from this remote ZIP file.
This is just a test to see if it works and if CORS policies
do work on the provided server.